rebase: reconcile native port against upstream doldecomp/sms (26 commits)
Rebased our 368 native-port commits onto upstream/main (was 26 commits ahead
at fork-point 0686c767; now on top of 4fa80205 "Mostly match SelectDir").

The mechanical rebase (-X theirs) left our and upstream's edits interleaved
hunk-by-hunk in files we both touched, producing internally-inconsistent TUs
(header renamed unkB4->getViewMtx / TLookAtCamera 2->6-arg / TMapCollisionBase
unk10->unk20 / TBathWaterManager layout / TMapXlu getter dropped, while our
.cpp still used the old API; plus duplicate class-member decls where both sides
added the same method). File-level resolution instead: for every file EITHER
side changed since the fork-point, take our known-good pre-rebase version
(our port wins on contested files); upstream's improvements land only on the
~200 files we never touched, plus genuinely-new upstream files (Animal/boid,
Animal/fishoid, Enemy/BossManta, Enemy/Igaiga headers, dolphin/gd/GDVert.h).

Added TVec3<f32> operator/ and operator/= (scalar divide) for upstream's new
Animal/boid.cpp. Builds clean (sms-boot links). Boot behavior is IDENTICAL to
the pre-rebase tip (both reach APP_STATE_GAMEPLAY at SB_STAGE=1 and both hit
the same pre-existing TWaterGun::changeBackup boot crash — NOT introduced by
this rebase; a separate port-HEAD regression to chase next).

+++ src/JSystem/JAudio/JAInterface/JAIGFrameStream.cpp
@@ -726,7 +726,7 @@ namespace StreamLib {
 
 		for (u32 i = 0; i < ARRAY_COUNT(assign_ch); ++i) {
 			if (assign_ch[i] != nullptr && assign_ch[i]->unk8 != 0)
-				JASystem::TDSPChannel::free(assign_ch[i], (u32)&assign_ch[i]);
+				JASystem::TDSPChannel::free(assign_ch[i], (u32)(uintptr_t)&assign_ch[i]);
 
 			assign_ch[i] = nullptr;
 		}
@@ -768,8 +768,8 @@ namespace StreamLib {
 		}
 
 		if (assign_ch[0] == nullptr) {
-			assign_ch[0] = JASystem::TDSPChannel::alloc(0, (u32)&assign_ch[0]);
-			assign_ch[1] = JASystem::TDSPChannel::alloc(0, (u32)&assign_ch[1]);
+			assign_ch[0] = JASystem::TDSPChannel::alloc(0, (u32)(uintptr_t)&assign_ch[0]);
+			assign_ch[1] = JASystem::TDSPChannel::alloc(0, (u32)(uintptr_t)&assign_ch[1]);
 			if (assign_ch[0] != nullptr && assign_ch[1] != nullptr) {
 				assign_ch[0]->unk3 = 0x7F;
 				assign_ch[1]->unk3 = 0x7F;
@@ -826,9 +826,9 @@ namespace StreamLib {
 			if (buf->unk2 != 0) {
 				if (adpcmbuf_state != 2) {
 					JASystem::TDSPChannel::free(assign_ch[0],
-					                            (u32)&assign_ch[0]);
+					                            (u32)(uintptr_t)&assign_ch[0]);
 					JASystem::TDSPChannel::free(assign_ch[1],
-					                            (u32)&assign_ch[1]);
+					                            (u32)(uintptr_t)&assign_ch[1]);
 					hackyHackMcHackson(-1);
 					playflag  = 0;
 					playflag2 = 2;
